Use snapshots and change domain
https://github.com/creek-service/creek-test/issues/24
https://github.com/creek-service/creek-test/issues/30
