+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,2 +1,3 @@
 ### Reviewer checklist
-- [ ] Documentation about what the example covers has been updated.
+- [ ] Ensure relevant issues are linked (description should include text like "Fixes #<issue number>")
+- [ ] Ensure any appropriate documentation has been added or amended
