+++ api/build.gradle.kts
@@ -7,7 +7,7 @@ val creekVersion : String by extra
 
 dependencies {
     api(project(":ids"))
-    api("org.creek:creek-kafka-metadata:$creekVersion")
+    api("org.creekservice:creek-kafka-metadata:$creekVersion")
 
     // To avoid dependency hell downstream, avoid adding any more dependencies except Creek metadata jars and test dependencies.
