+++ common/build.gradle.kts
@@ -8,6 +8,6 @@ val creekVersion : String by extra
 dependencies {
     api(project(":services"))
 
-    api("org.creek:creek-service-context:$creekVersion")
-    api("org.creek:creek-kafka-streams-extension:$creekVersion")
+    api("org.creekservice:creek-service-context:$creekVersion")
+    api("org.creekservice:creek-kafka-streams-extension:$creekVersion")
 }
