+++ word-counter-service/build.gradle.kts
@@ -7,5 +7,5 @@ val creekVersion : String by extra
 dependencies {
     implementation(project(":common"))
 
-    testImplementation("org.creek:creek-kafka-streams-test:$creekVersion")
+    testImplementation("org.creekservice:creek-kafka-streams-test:$creekVersion")
 }
